fix: fix improper data file handling
Move common data files (wordlists, headers, payloads, jwks locations) into a dedicated `files/` subdirectory within the package.

Update the code and packaging configuration to use `importlib_resources` to access these files correctly when the tool is installed as a package. This improves reliability by ensuring data files are found regardless of the installation method or current working directory.
