+++ mise.toml
@@ -1,0 +1,2 @@
+[tools]
+python = "3.11"
